refactor(types): extract IIndirectFire* into IndirectFireInterfaces.ts
Move 8 types (IndirectFireBasis, WeaponFireMode, IIndirectFireResolution,
IIndirectFireEventBase and the 4 payload variants) from CombatInterfaces.ts
(956 LOC) into a new IndirectFireInterfaces.ts (~149 LOC).

CombatInterfaces.ts shrinks to 824 LOC, pre-empting the 1000+ threshold
before Wave 9 aero-combat additions land.
